+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2023-07-29
+
+### Added
+
+- Added 'display videos' which gives basic info about linked video (from streamable/youtube) directly in the chat
+
+### Changed
+
+- Changed settings-manager to be more bullet proof
+
 ## [1.0.0] - 2023-07-20
 
 ### Added